+++ CHANGELOG.md
@@ -24,6 +24,7 @@ Please file a bug if you notice a violation of semantic versioning.
 ### Deprecated
 ### Removed
 ### Fixed
+- reduce extra leading whitespace in info table column 2
 ### Security
 
 ## [1.1.0] - 2025-09-02